fix: changed email id type

+++ email-api/src/main/java/com/gbtec/email/api/client/business/model/EmailDTO.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 import java.util.List;
 
 @Builder
-public record EmailDTO(Integer emailId, String emailFrom, List<String> emailTo, List<String> emailCC, String emailBody, int state) {
+public record EmailDTO(Long emailId, String emailFrom, List<String> emailTo, List<String> emailCC, String emailBody, int state) {
 }
